# EEDU-Backend/src/main/resources/ElementEDU-openapi.yaml
openapi: "3.0.3"
info:
  title: "ElementEDU API"
  description: "ElementEDU API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
  /user/login/twofactor/create:
    post:
      summary: "POST user/login/twofactor/create"
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCreateModel"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TwoFactorModel"
  /user/login/twofactor/enable/{method}/{code}:
    get:
      summary: "GET user/login/twofactor/enable/{method}/{code}"
      operationId: "enable"
      parameters:
        - name: "method"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "EMAIL"
              - "SMS"
              - "TOTP"
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
  /user/login/twofactor/select/{method}:
    get:
      summary: "GET user/login/twofactor/select/{method}"
      operationId: "select"
      parameters:
        - name: "method"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "EMAIL"
              - "SMS"
              - "TOTP"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
  /user/login/twofactor/verify/{code}:
    get:
      summary: "GET user/login/twofactor/verify/{code}"
      operationId: "verify"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
components:
  securitySchemes: # Define your security scheme here
    BearerAuthentication: # Define 'BearerAuth', to be referenced by operations
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TwoFactorModel:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        method:
          type: "string"
          enum:
            - "EMAIL"
            - "SMS"
            - "TOTP"
        enabled:
          type: "boolean"
        claims:
          type: "string"
      required:
        - "id"
        - "method"
...
    TwoFactorCreateModel:
      type: "object"
      properties:
        userID:
          type: "integer"
          format: "int64"
        method:
          type: "string"
        data:
          type: "string"
      required:
        - "userID"
        - "method"
        - "data"
...